backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "aiofiles==23.2.1",
    "fastapi==0.104.1",
    "google-generativeai==0.3.2",
    "httpx==0.25.2",
    "openai==1.3.7",
    "passlib[bcrypt]==1.7.4",
    "pydantic==2.5.0",
    "python-dateutil==2.8.2",
    "python-dotenv==1.0.0",
    "python-jose[cryptography]==3.3.0",
    "python-multipart==0.0.6",
    "requests==2.31.0",
    "uvicorn[standard]==0.24.0",
]
